backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
mysqlclient>=2.2.0
django-cors-headers>=4.3.0
Pillow>=10.2.0
